+++ src/pd_esp32/esp32_queue.cpp
@@ -543,4 +543,18 @@ void fas_init_engine(FastAccelStepperEngine* engine, uint8_t cpu_core) {
   }
 }
 
+AqeResultCode FastAccelStepperEngine::synchronizedStart(
+    FastAccelStepper** const steppers, uint8_t cnt) {
+  AqeResultCode rc = AqeResultCode::OK;
+  fasDisableInterrupts();
+  for (uint8_t i = 0; i < cnt; i++) {
+    AqeResultCode e = steppers[i]->addQueueEntry(NULL, true);
+    if (rc == AqeResultCode::OK && e != AqeResultCode::OK) {
+      rc = e;
+    }
+  }
+  fasEnableInterrupts();
+  return rc;
+}
+
 #endif