+++ src/pd_pico/pico_queue.cpp
@@ -296,4 +296,19 @@ void fas_init_engine(FastAccelStepperEngine* engine) {
 
   program = stepper_make_program();
 }
+
+AqeResultCode FastAccelStepperEngine::synchronizedStart(
+    FastAccelStepper** const steppers, uint8_t cnt) {
+  AqeResultCode rc = AqeResultCode::OK;
+  fasDisableInterrupts();
+  for (uint8_t i = 0; i < cnt; i++) {
+    AqeResultCode e = steppers[i]->addQueueEntry(NULL, true);
+    if (rc == AqeResultCode::OK && e != AqeResultCode::OK) {
+      rc = e;
+    }
+  }
+  fasEnableInterrupts();
+  return rc;
+}
+
 #endif